adding sql file

+++ src/main/java/com/edanrh/apiong/repository/RepositoryRole.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.edanrh.apiong.repository.entities.RoleEntity;
 
-public interface RepositoryRole extends CrudRepository<RoleEntity, BigInteger>{
+public interface RepositoryRole extends CrudRepository<RoleEntity, Long>{
     
 }
