+++ src/main/java/com/edanrh/apiong/repository/RepositoryUser.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.edanrh.apiong.repository.entities.UserEntity;
 
-public interface RepositoryUser extends CrudRepository<UserEntity,BigInteger> {
+public interface RepositoryUser extends CrudRepository<UserEntity,Long> {
 
     Optional<UserEntity> findByEmail(String email);
     Optional<UserEntity> deleteByEmail(String email);
